requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
openai
pymupdf             # fitz (extrair texto de PDFs)
pytesseract         # OCR fallback (requer tesseract instalado no SO)
Pillow              # imagens para OCR
qdrant-client       # cliente Qdrant (opcional - se você usar Qdrant)
sentence-transformers  # embeddings local fallback
scikit-learn        # NearestNeighbors fallback
numpy